removing redundant classes
