+++ pydp/algorithms/laplacian/_count.py
@@ -1,5 +1,10 @@
-from .._algorithm import Algorithm
+from .._algorithm import MetaAlgorithm
 
 
-class Count(Algorithm):
-    pass
+class Count(MetaAlgorithm):
+    '''
+    Count Explaination
+    TODO
+    '''
+    def __init__(self, epsilon: float=1.0, dtype: str="int"):
+        super().__init__(epsilon=epsilon, dtype=dtype)
