+++ pydp/algorithms/laplacian/_percentile.py
@@ -1,5 +1,27 @@
-from .._algorithm import BoundedAlgorithm, PercentileBase
+from .._algorithm import MetaAlgorithm
+from typing import Union
 
 
-class Percentile(PercentileBase):
-    pass
+class Percentile(MetaAlgorithm):
+    def __init__(
+        self,
+        epsilon: float=1.0,
+        percentile: float=0.0,
+        lower_bound: Union[int, float, None]=None,
+        upper_bound: Union[int, float, None]=None,
+        dtype:str="int",
+    ):
+        super().__init__(
+            epsilon=epsilon,
+            percentile=percentile,
+            lower_bound=lower_bound,
+            upper_bound=upper_bound,
+            dtype=dtype,
+        )
+
+    @property
+    def percentile(self)-> float:
+        """
+        percentile Gets the value that was set in the constructor.
+        """
+        return self._MetaAlgorithm__algorithm.percentile
